Implemented creating, editing, and updating itineraries
- Updated version of NodeJS in Vagrant provisioning
- Downloading routes, tracks and waypoints
- Return HTTP Status 500 if an exception is caught when handling request
- Show the main menu items that have not yet been implemented as disabled
- Upgrade to Bootstrap 5.2.2
- Bug fix - Location tracking notes stored as empty string instead of NULL
- Bug fix - Show point on map when link clicked in tracked location list

+++ src/geo_utils.cpp
@@ -26,7 +26,7 @@
 using namespace fdsd::trip;
 using json = nlohmann::basic_json<nlohmann::ordered_map>;
 
-location::location(long long id,
+location::location(long id,
                    double lon,
                    double lat,
                    std::pair<bool, double> altitude)
